refactor(b): the shape arithmetic leaves the reading, and the sentence reads as one
Decomplect, and a clarity pass over what is left.

Two different kinds of thing were braided into `narrowing.ts`. One is a
READING — a memo graph over what the reader typed, what the store
published and what the clock says, which is about TIME. The other is
arithmetic over a VALUE: given a page and a set of ids, which rows
survive, how many places there are, how many of them the query selected,
and whether anything on it was put away. The second is pure, total over
`Drawn`, and knows nothing about a query, a preference or a signal — so
it is now `filter/drawn.ts`, and the reading imports four functions.

What the split buys beyond tidiness: every function in the new file is
one arm per page shape, and read beside five paragraphs of memo argument
that table was invisible — while it is exactly what has to grow an arm
the day a sixth kind of page draws nodes. HACKING.md's rule, and its
reason: files and a hierarchy over a module holding everything about one
word.

The clarity half: `countLine`'s two conditionals are named where they
are decided ("more" and the plural) rather than nested inside the
template they interpolate into, so the function reads as the sentence it
builds; the doc links that crossed the new seam now say which file they
cross to.

No behaviour changes.
